+++ crates/symbolicator/src/endpoints/proxy.rs
@@ -24,7 +24,7 @@ async fn load_object(state: Service, path: String) -> anyhow::Result<Option<Arc<
         None => return Ok(None),
     };
 
-    log::debug!("Searching for {:?} ({:?})", object_id, filetypes);
+    tracing::debug!("Searching for {:?} ({:?})", object_id, filetypes);
 
     let found_object = state
         .objects()