+++ crates/symbolicator/src/services/download/sentry.rs
@@ -139,10 +139,10 @@ impl SentryDownloader {
         let response = request.send().await?;
 
         if response.status().is_success() {
-            log::trace!("Success fetching index from Sentry");
+            tracing::trace!("Success fetching index from Sentry");
             Ok(response.json().await?)
         } else {
-            log::warn!("Sentry returned status code {}", response.status());
+            tracing::warn!("Sentry returned status code {}", response.status());
             Err(SentryError::BadStatusCode(response.status()))
         }
     }
@@ -172,7 +172,7 @@ impl SentryDownloader {
             }
         }
 
-        log::debug!(
+        tracing::debug!(
             "Fetching list of Sentry debug files from {}",
             &query.index_url
         );
@@ -274,7 +274,7 @@ impl SentryDownloader {
         match request.await {
             Ok(Ok(response)) => {
                 if response.status().is_success() {
-                    log::trace!("Success hitting {}", download_url);
+                    tracing::trace!("Success hitting {}", download_url);
 
                     let content_length = response
                         .headers()
@@ -291,17 +291,17 @@ impl SentryDownloader {
                     response.status(),
                     StatusCode::FORBIDDEN | StatusCode::UNAUTHORIZED
                 ) {
-                    log::debug!("Insufficient permissions to download from {}", download_url);
+                    tracing::debug!("Insufficient permissions to download from {}", download_url);
                     Err(DownloadError::Permissions)
                 } else if response.status().is_client_error() {
-                    log::debug!(
+                    tracing::debug!(
                         "Unexpected client error status code from {}: {}",
                         download_url,
                         response.status()
                     );
                     Ok(DownloadStatus::NotFound)
                 } else {
-                    log::debug!(
+                    tracing::debug!(
                         "Unexpected status code from {}: {}",
                         download_url,
                         response.status()
@@ -310,7 +310,7 @@ impl SentryDownloader {
                 }
             }
             Ok(Err(e)) => {
-                log::debug!("Skipping response from {}: {}", download_url, e);
+                tracing::debug!("Skipping response from {}: {}", download_url, e);
                 Err(DownloadError::Reqwest(e)) // must be wrong type
             }
             // Timed out