+++ crates/symbolicator/src/test.rs
@@ -21,9 +21,10 @@ use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 
-use log::LevelFilter;
 use reqwest::Url;
 use serde::{Deserialize, Serialize};
+use tracing_subscriber::filter::EnvFilter;
+use tracing_subscriber::fmt::fmt;
 use warp::filters::fs::File;
 use warp::reject::{Reject, Rejection};
 use warp::Filter;
@@ -43,9 +44,11 @@ pub use tempfile::TempDir;
 ///  - Initializes logs: The logger only captures logs from the `symbolicator` crate and mutes all
 ///    other logs (such as actix or symbolic).
 pub(crate) fn setup() {
-    env_logger::builder()
-        .filter(Some("symbolicator"), LevelFilter::Trace)
-        .is_test(true)
+    fmt()
+        .with_env_filter(EnvFilter::new("symbolicator=trace"))
+        .with_target(false)
+        .pretty()
+        .with_test_writer()
         .try_init()
         .ok();
 }